Refactor: Main module
